Server/requirements.txt
Flask>=2.2.5
flask-cors>=4.0.0
PyJWT>=2.8.0
crewai>=0.1.0
langchain>=0.1.0
langchain-groq>=0.0.5
langchain-openai>=0.0.5
python-dotenv>=1.0.0
reportlab>=4.0.0
requests>=2.31.0
gunicorn>=20.1.0
matplotlib>=3.8.0
